+++ src/Microsoft.Azure.Functions.Extensions.Mcp/Diagnostics/TraceConstants.cs
@@ -1,0 +1,12 @@
+﻿namespace Microsoft.Azure.Functions.Extensions.Mcp.Diagnostics;
+
+internal sealed class TraceConstants
+{
+    public const string ExtensionActivitySource = "Azure.Functions.Extensions.Mcp";
+    public const string ExtensionActivitySourceVersion = "1.0.0.0";
+
+    public const string ExceptionEventNameAttribute = "exception";
+    public const string ExceptionTypeAttribute = "exception.type";
+    public const string ExceptionMessageAttribute = "exception.message";
+    public const string ExceptionStacktraceAttribute = "exception.stacktrace";
+}