+++ src/release_notes.md
@@ -4,8 +4,8 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Extensions 1.0.0-preview.5
+### Microsoft.Azure.Functions.Extensions 1.0.0-preview.7
 
-- Restricting SSE endpoint to GET requests.
+- Adding instrumentation to emit a server span for the tools/call. (#47)
 
-### Microsoft.Azure.Functions.Worker.Extensions.Mcp 1.0.0-preview.5
+### Microsoft.Azure.Functions.Worker.Extensions.Mcp 1.0.0-preview.7